+++ src/RestInPatches/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## 0.1.6 | 5 September 2026
+## 0.1.6 | 13 September 2026
 
 - Fixed the Max Footprints setting having no effect
 - Footprints now fade away instead of vanishing on the spot
+- Added a setting for how quickly footprints fade out when the limit removes them
+- Fixed the rest of the mod failing to load when another footprint mod is installed
 
 ## 0.1.5 | 20 June 2026